+++ src/main/java/com/skycat/mystical/common/spell/cure/SpellCure.java
@@ -11,6 +11,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
 
+@SuppressWarnings("rawtypes")
 public abstract class SpellCure {
     @Getter protected int contributionGoal;
     @Getter protected final Class cureType;
@@ -60,6 +61,9 @@ public abstract class SpellCure {
     }
 
     public void contribute(@Nullable UUID uuid, int amount) {
+        if (amount == 0) {
+            return;
+        }
         if (uuid == null) {
             uuid = UUID.fromString("09b4c37c-1dd6-4eb4-8adf-f660dd111410"); // This better not collide with anything.
         }
@@ -79,30 +83,16 @@ public abstract class SpellCure {
         return contributionTotal >= contributionGoal;
     }
 
-    /**
-     * Award power to players based on their contributions.
-     * @param totalPower The total amount of power to distribute among the players
-     * @implNote Simple overload of {@link #awardPower(int, int)}.
-     */
-    public void awardPower(int totalPower) {
-        awardPower(totalPower, Integer.MAX_VALUE);
-    }
-
-    /**
-     * Award power to players based on their contributions.
-     * @param totalPower The total power to distribute among contributors.
-     * @param max The maximum power to give to any one contributor.
-     */
-    public void awardPower(int totalPower, int max) {
-        for (UUID uuid : contributions.keySet()) {
-            if (contributions.get(uuid) <= 0) continue;
-            // Formula: min(totalPower * percentContributed, max)
-            Mystical.getHavenManager().addPower(uuid, (int) Math.min(totalPower * ((double) contributions.get(uuid) / contributionTotal), max));
-        }
-    }
-
     public int getContributionsLeft() {
         return getContributionGoal() - getContributionTotal();
+    }
+
+    public int getContributionsOf(UUID contributor) {
+        return contributions.get(contributor);
+    }
+
+    public int getContributorCount() {
+        return contributions.size();
     }
 
 }